Move paraseBaseMediaDecodeTime to mp4_segment_parsers per code review suggestion.
