+++ lib/media/timestamp_offset_corrector.js
@@ -10,8 +10,6 @@ goog.require('shaka.log');
 goog.require('shaka.media.SegmentReference');
 goog.require('shaka.util.FakeEvent');
 goog.require('shaka.util.ManifestParserUtils');
-goog.require('shaka.util.Mp4BoxParsers');
-goog.require('shaka.util.Mp4Parser');
 goog.require('shaka.util.Mp4SegmentParsers');
 
 /**
@@ -126,7 +124,9 @@ shaka.media.TimestampOffsetCorrector = class {
     }
 
     const segmentStartTime =
-        this.parseBaseMediaDecodeTime_(mediaData, segmentData) +
+        shaka.util.Mp4SegmentParsers.parseBaseMediaDecodeTime(
+            segmentData, 
+            mediaData.trackTimescales) +
         reference.timestampOffset;
     const timestampDiscrepancy =
         segmentStartTime - reference.getStartTime();
@@ -161,57 +161,6 @@ shaka.media.TimestampOffsetCorrector = class {
     mediaData.timestampOffset = correctedTimestampOffset;
     return timestampOffsetChanged;
   }
-
-  /**
-   * Parses the baseMediaDecodeTime from a media segment and
-   * converts it to seconds using timescale from the
-   * associated init segment.
-   *
-   * @param {!shaka.media.TimestampOffsetCorrector.MediaData_} mediaData
-   * @param {BufferSource} segmentData
-   * @return {number} base media decode time in seconds
-   * @private
-   */
-  parseBaseMediaDecodeTime_(mediaData, segmentData) {
-    const Mp4Parser = shaka.util.Mp4Parser;
-
-    // Fields that are found in MOOF boxes
-    let baseMediaDecodeTime = 0;
-    let trackId = 0;
-
-    new Mp4Parser()
-        .box('moof', Mp4Parser.children)
-        .box('traf', Mp4Parser.children)
-        .fullBox('tfhd', (box) => {
-          if (box.flags !== null) {
-            const parsedTFHD = shaka.util.Mp4BoxParsers.parseTFHD(
-                box.reader, box.flags);
-            trackId = parsedTFHD.trackId;
-          }
-        })
-        .fullBox('tfdt', (box) => {
-          if (box.version !== null) {
-            const parsedTFDT = shaka.util.Mp4BoxParsers.parseTFDT(
-                box.reader, box.version);
-            baseMediaDecodeTime = parsedTFDT.baseMediaDecodeTime;
-          }
-        })
-        .parse(segmentData, /* partialOkay= */ false);
-
-    let timescale = 1;
-    if (mediaData.trackTimescales.has(trackId)) {
-      timescale = mediaData.trackTimescales.get(trackId);
-    }
-    const baseMediaDecodeTimeSec = baseMediaDecodeTime / timescale;
-    shaka.log.v2('parseBaseMediaDecodeTime_()',
-        'trackId', trackId,
-        'timescale', timescale,
-        'baseMediaDecodeTime', baseMediaDecodeTime,
-        'baseMediaDecodeTimeSec', baseMediaDecodeTimeSec);
-
-    return baseMediaDecodeTimeSec;
-  }
-
 
   /**
    * Parses timescales from an init segment
