+++ lib/util/mp4_segment_parsers.js
@@ -7,6 +7,7 @@
 goog.provide('shaka.util.Mp4SegmentParsers');
 
 goog.require('goog.asserts');
+goog.require('shaka.log');
 goog.require('shaka.util.Mp4Parser');
 goog.require('shaka.util.Mp4BoxParsers');
 
@@ -67,5 +68,54 @@ shaka.util.Mp4SegmentParsers = class {
     });
 
     return trackIdToTimescale;
+  }
+
+  /**
+   * Parses the baseMediaDecodeTime from a media segment and converts 
+   * to seconds using trackTimescale, that may be obtained by a previous
+   * call to `parseTrackTimescales`.
+   *
+   * @param {BufferSource} segmentData
+   * @param {Map<number,number>} trackTimescales
+   * @return {number} base media decode time in seconds
+   */
+   static parseBaseMediaDecodeTime(segmentData, trackTimescales) {
+    const Mp4Parser = shaka.util.Mp4Parser;
+
+    // Fields that are found in MOOF boxes
+    let baseMediaDecodeTime = 0;
+    let trackId = 0;
+
+    new Mp4Parser()
+        .box('moof', Mp4Parser.children)
+        .box('traf', Mp4Parser.children)
+        .fullBox('tfhd', (box) => {
+          if (box.flags !== null) {
+            const parsedTFHD = shaka.util.Mp4BoxParsers.parseTFHD(
+                box.reader, box.flags);
+            trackId = parsedTFHD.trackId;
+          }
+        })
+        .fullBox('tfdt', (box) => {
+          if (box.version !== null) {
+            const parsedTFDT = shaka.util.Mp4BoxParsers.parseTFDT(
+                box.reader, box.version);
+            baseMediaDecodeTime = parsedTFDT.baseMediaDecodeTime;
+          }
+        })
+        .parse(segmentData, /* partialOkay= */ false);
+
+    let timescale = 1;
+    if (trackTimescales.has(trackId)) {
+      timescale = trackTimescales.get(trackId);
+    }
+    const baseMediaDecodeTimeSec = baseMediaDecodeTime / timescale;
+    shaka.log.v2('parseBaseMediaDecodeTime()',
+        'trackId', trackId,
+        'timescale', timescale,
+        'baseMediaDecodeTime', baseMediaDecodeTime,
+        'baseMediaDecodeTimeSec', baseMediaDecodeTimeSec);
+
+    return baseMediaDecodeTimeSec;
   }
 };
